Refeat- (Front/Detail) boton de confirmacion de eliminar

+++ Client/src/pages/Detail/Detail.jsx
@@ -57,20 +57,28 @@ const Detail = () => {
     setModalProduct(!modalProduct);
   };
 
-  useEffect(() => {
-    if (msg.length > 2) {
-      Swal.fire({
-        title: "SUCCESS!",
-        text: msg,
-        icon: "success",
-      });
-      navigate("/products");
-    }
-  }, [product]);
-
   const handleDeleteProduct = async () => {
-    await dispatch(deleteProduct(product[0]?.PRODUCTO?.id_producto, branch));
-    setProducts({});
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "¡No podrás revertir esto!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, borrarlo",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+        });
+        // Si el usuario ha confirmado, ejecuta la acción de borrar el producto
+        dispatch(deleteProduct(product[0]?.PRODUCTO?.id_producto, branch));
+        setProducts({});
+        navigate("/products");
+      }
+    });
   };
 
   return (
